TS: Extracting EntryCard from Enrtry page
