+++ src/components/Entry/Card/styles.css
@@ -1,0 +1,9 @@
+.md-small {
+  display: flex;
+  width: 46px;
+}
+
+.new-card-description {
+  overflow: hidden;
+  height: 6em; /*better for alignment description + counter*/
+}
